Implement describe and compare logic
In places the describe logic is looking slightly further ahead to when
human-readable representations need to be formatted.

The compare logic allows for symbols, functions and promises requiring
looser comparisons after the descriptors have been deserialized from a
snapshot.

This breaks compatibility with `lodash.isequal` in several ways:

* Primitives no longer compare to wrapped primitives (`1 versus
Object(1)`)
* The Arguments object must be compared to an array, and this comparison
only works when the Arguments object is in the `actual` value
* Array non-index properties are compared (indeed there is no special
handling of Arrays)
* Map and Set entries are compared in-order
* Symbol properties are compared in-order (unlike property names).
These properties cannot be sorted, and comparing them in an unordered
fashion is more complicated