Potentially improve getting object keys
Calling code already knows whether list item accessors should be
rejected. There's no need to check if `obj` has a length property.

We can safely set the length to 0 when list item accessors are not
excluded, which has the advantage of the variable always having the same
type, making the function easier to optimize.