Fix multiline string value diffs of properties and map entries
When diffing strings:

* return a formatted diff for each line
* track whether the line is the last for the actual or expected string
* add a property to identify neutral lines

Explicitly ensure that when formatting diffs for properties and map
entries, `formatDiff()` is only used if the value is indeed a string.
Add comments explaining the logic further.

Correctly prefix the first actual or neutral, and expected lines with
the property or map entry key. Correctly add a comma after the last
actual and expected, or neutral, line in the string.