Support themes
Fixes #7.

* Enhance formatting logic so more output can be themed

* Don't be clever and use backticks when a string contains straight
quotes. It's trickier when using themes. Allow to-be-escaped-quote for
single line strings to be configured in the theme, but hardcode the
assumption that multiline strings use backticks

* Refactor and improve string diffs. Replace hardcoded underline ansi
style with theme support. Allow control pictures to be themed

* Fix formatting of date objects with extra properties

* Fix error formatting: actually print `ctor` if `name` is falsy, only
print string tag if it does not match the `ctor`

* In diffs, format stats with a neutral gutter. The stats separator is
shown because something else is different, so there's no need to draw
attention to it by showing a gutter