Improve handling of circular references
* Detect circular references in comparisons. Circular references are
detected once pointers are looked up. If both sides have circular
references whose original pointers were not equal, treat the sides as
unequal.

* Pass an `isCircular()` to `compareDiff()` and `rebalanceDiff()`
methods. This is useful for items, map entries and properties which need
to compare as unequal if they contain a circular reference. Otherwise,
when their values are compared and are deemed circular, they won't be
able to be formatted correctly

* Subclass `Set` to create a `Circular` class which excludes items, map
entries and properties from being tracked. Since these are container
values they'll never be circular