Modify themes when formatting descriptors
This is useful for plugins which may want to use a built-in descriptor,
but override the theme used during formatting. Due to the stack-based
format() and diff() implementation, modifications will not leak while
recursing. That said, it's possible for descriptors to deeply format
another primitive, and they can do that with the original, unmodified
theme.