Refactor formatting and diffing
Split formatting into a `formatDeep()` and a `formatShallow()`.
Descriptors should implement at least one of them. `formatDeep()` is
expected to be used when there is no `createRecursor()` implementation.

Optionally descriptors can implement a `diffDeep()` and `diffShallow()`.
`diffDeep()` should only be necessary if differences can't be determined
through recursion. `diffShallow()` may be useful if descriptors are
deemed shallowly equal for diffing purposes, even though differences
do exist that cannot be determined through recursion.

Refactor `rebalanceDiff()` into a `prepareDiff()`. This allowed
`compareDif()` to removed, which makes the diffing algorithm easier to
understand, and enables descriptors to influence the algorithm.

Introduce a `lineBuilder` and build lines, rather than concatenating
strings. Flag which lines are first and last and keep this accurate as
stacks unwind. This makes it easier to fix indentation, add separators
and merge different line groups.

The `lineBuilder` can build lines and collections of lines. Both
implement the same API, so formatting code doesn't need to check what
it received. Lines and collections are mostly immutable.

String diffing has been refactored to use the line builder, which
greatly increases readability of the algorithms.

Provide an indenter to the formatting & diffing methods, allowing those
methods to increase the indent level if necessary.

Remove `extraneous` and `missing` gutter types. Simplify in favor of
`actual` and `expected`, that is, the origin of the value. Rename gutter
themes.

Tighten theme usage to make sure themes like `string.open` and
`string.close` don't span across different lines. Reserve open/close
pairs for values that are shown on a single line. Rename properties that
span multiple lines.

Track original descriptors so that formatters receive an `origin` that
they emitted through their recursors. This helps the `mapEntry`
descriptor when keys are pointers.