Add option to invert diffs
When Concordance is used for snapshot testing the `actual` (left-hand
side) value should be the deserialized snapshot. This can make the diff
result confusing since the user might reasonably presume the snapshot
value is the *expected* value, with a `+` gutter, not a `-` gutter.

This commit adds an option to invert diffs, reordering diff lines and
changing the gutters so it looks like the snapshot value was indeed
the *expected* value.