Simplify pointer implementation
Before this commit, Concordance attempted to determine whether a complex
value was reused in the same location on either side of a comparison. It
did this by detecting value reuse and emitting a Pointer descriptor
instead. This required lookups to be maintained, to map pointers back to
previously seen descriptors. It turned out that maintaining these
lookups is hard, and there were various code paths where Concordance
should have added descriptors to the lookup tables but didn't.
Furthermore, pointer indexes are simple integer increments, which means
that the same value in either side of the comparison may actually have a
different index. This breaks the comparison logic.

https://github.com/avajs/ava/issues/1431 shows some of the issues that
arose from this approach.

This commit simplifies the implementation. When describing, if the same
value is encountered then the first descriptor is emitted. The
serialization logic tracks complex descriptors by their pointer index
and serializes a pointer descriptor if reuse is detected.
Deserialization ensures that the first descriptor is emitted, rather
than the pointer.

Circular references are now tracked for each side of the comparison. A
stack is maintained, and if at any stage in the comparison a circular
reference is detected, then the result of that comparison is only equal
if both sides have the same circular reference. Which is a long way of
saying that some circular references are considered equal, and others
are not. This shouldn't matter much for real-life uses of Concordance,
though it is a breaking change.