Implement flags to control behavior
Flags can be provided when comparing, describing and deserializing.

Library users can derive their own flags from a default set of flags in their disabled states. This is the recommended
way to use flags, since any new behavior won't be opted in to by default.

However, by default, the library will use a set of flags which may not be in their disabled state.

Arguments-to-array comparisons are now behind a flag, and disabled by defaault.