Simplify symbol property serialization
Rely on fullyDeserialize() to proactively deserialize the property value. The value no longer needs to be encoded as a
uint8array, which was required so a single read (of the uint8array) would progress the decoder to the next property.

The downside is that deeply nested property values are fully deserialized, which could be unnecessary when an object has
a single symbol property and is compared against another object with a different symbol property. However in more
typical scenarios this shouldn't have a measurable performance impact, if only because if values are equal we'll
eventually fully deserialize anyway.