Give RHS control over being compared
Most complex values extend `ObjectRepresentation`. If LHS in a comparison is an `ObjectRepresentation`, it could be
compared against any other subclass.

Normally constructor names won't match, but that's not the case for fuzzy matching.

This adds a new `acceptsComparisonFrom` method to all value represenations. It enables RHS to control whether it can be
compared to LHS. Subclasses are expected to override to avoid unintentional comparisons from their parent class.

This also allows us to invert control of certain exceptional comparison scenarios, such as whether an arguments object
can be compared to an array, or a sparse array value to `undefined`, plus a host of fuzzy matching behaviors.

Consequently we can also remove most static `is()` methods, since the various representation classes don't need to
reference each other.